# app/dist/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.0",
        "title": "Apartment Butler API"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
...
        "/user/{userId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "description": "This endpoint returns information about the user by its id.\n",
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "User Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "description": "This endpoint updates information of the user by its id.\n",
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "User Access Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "firstName",
                        "in": "formData",
                        "description": "First name of User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastName",
                        "in": "formData",
                        "description": "Last name of User",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An Object of User",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer",
                    "format": "int32"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                }
            }
        },
...
    }
}